Improved memory efficiency

+++ contracts/Market.sol
@@ -20,9 +20,9 @@ contract Market {
 
     address[] public products;
     
-    event ProductCreated(string name, uint quantity, uint price, uint currency, uint[] _region, uint[] _category);
+    event ProductCreated(string name, uint32 quantity, uint price, uint8 currency, uint8[] _region, uint8[] _category);
 
-    function createProduct (string memory _name, uint _quantity, uint _price, uint _currency, uint[] memory _region, uint[] memory _category) public returns (address) {
+    function createProduct (string memory _name, uint32 _quantity, uint _price, uint8 _currency, uint8[] memory _region, uint8[] memory _category) public returns (address) {
         require (msg.sender == owner, "Caller does not own Market");
         
         Product product = new Product(payable(msg.sender), _name, _quantity, _price, _currency, _region, _category);
